# remove ring buffer usage from double_buffered_gemm