Usage of `ring_buffer::smem_ring_buffer`  was causing spills to local memory even though the compiler had abundant register budget to avoid the same. I suspect this has something to do with class variable visibility or something to that affect.
Though I would like to make use of the ring buffer class later on as the buffer depth increases with async copies, I have kept the file, but removed it's usage replacing it with a simple index counter for the same. 

difference in performance on a 3060 Max Q:

| Size      | Main |   This PR |
| ----------- | ----------- |---------|
| 1024 x 1024 x 1024      | 2.435 TFlops       |       3.43 TFLops|
| 2048 x 2048 x 2048   | 3.9 TFlops        |      5.538 TFlops|
|4096 x 4096 x 4096    | 3.9 TFLops       |   5.532 TFLops                          |